+++ src/Codec/Archive/Foreign/Raw.hs
@@ -1,0 +1,6 @@
+module Codec.Archive.Foreign.Raw ( module Codec.Archive.Foreign.Archive.Raw
+                                 , module Codec.Archive.Foreign.ArchiveEntry.Raw
+                                 ) where
+
+import           Codec.Archive.Foreign.Archive.Raw
+import           Codec.Archive.Foreign.ArchiveEntry.Raw